requirements.txt
fastapi>=0.104.1,<1.0.0
uvicorn[standard]==0.40.0
pydantic>=2.12.5,<3.0.0
pydantic-settings>=2.12.0,<3.0.0
transformers>=4.35.2
torch==2.6.0
python-dotenv==1.2.1
pytest>=7.4.3
httpx>=0.25.2
google-cloud-storage
google-cloud-firestore
